feature:add multiple messages with same reply, handling all file types in reply , fixin bugs in middleware

+++ middleware/middleware.upload.js
@@ -4,41 +4,42 @@ import fs from "fs/promises"; // Import fs.promises instead of node:fs/promises
 import replyModel from "../Database/Schema/replySchema.js";
 
 let fileName;
-
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, path.join(path.resolve(
-        "../whatsappbot/media"
-      )));
-  },
-  filename: function (req, file, cb) {
-    fileName =  file.originalname;
-    cb(null, fileName);
-  },
-});
-const uploadMedia = multer({ storage }).single("media");
-
-const uploadingMediaMW =  (req, res, next) => {
+let fileNameinstance;
+const uploadingMediaMW = (req, res, next) => {
+  const storage = multer.diskStorage({
+    destination: function (req, file, cb) {
+      cb(null, path.join(path.resolve("./media")));
+    },
+    filename: function (req, file, cb) {
+      fileName = file.originalname;
+      cb(null, fileName);
+    },
+  });
+  const uploadMedia = multer({ storage }).single("media");
   const newUploadMediaInstance = uploadMedia;
   newUploadMediaInstance(req, res, async function (err) {
-    try {
-      console.log({ fileName });
-      const fileFounded = await replyModel.findOne({message:fileName})
-      console.log(fileFounded);
-     if (fileFounded){
-       const filePath = path.resolve(`../whatsappbot/media/${fileName}`);
-       console.log(filePath);
-       await fs.unlink(filePath); 
-       console.log("Successfully deleted filePath");
-      }} catch (err) {
+    if (fileName) {
+      try {
+        fileNameinstance = fileName;
+        const fileFounded = await replyModel.findOne({ message: fileName });
+        if (fileFounded) {
+          const filePath = path.resolve(`./media/${fileName}`);
+          await fs.unlink(filePath);
+          console.log("Successfully deleted ");
+        }
+      } catch (err) {
         console.error("Error deleting file:", err);
-      
+      }
+      if (err instanceof multer.MulterError) {
+        return res.status(500).json({ error: err.message });
+      }
+      fileName = "";
+      next();
+    } else {
+      fileNameinstance = "";
+      next();
     }
-    if (err instanceof multer.MulterError) {
-      return res.status(500).json({ error: err.message });
-    }
-    next();
+    console.log({ fileName });
   });
 };
-
-export { uploadingMediaMW, fileName };
+export { uploadingMediaMW, fileNameinstance };
